+++ src/DebugBar/Resources/widgets.js
@@ -39,6 +39,16 @@ if (typeof(PhpDebugBar) == 'undefined') {
         return value;
     };
 
+    /**
+     * Returns a prefixed css class name
+     * 
+     * @param {String} cls
+     * @return {String}
+     */
+    var csscls = function(cls) {
+        return PhpDebugBar.utils.csscls(cls, 'phpdebugbar-widgets-');
+    };
+
 
     // ------------------------------------------------------------------
     // Generic widgets
@@ -55,7 +65,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
 
         tagName: 'ul',
 
-        className: 'phpdebugbar-widgets-list',
+        className: csscls('list'),
 
         initialize: function(options) {
             if (!options['itemRenderer']) {
@@ -73,7 +83,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
 
                 var data = this.get('data');
                 for (var i = 0; i < data.length; i++) {
-                    var li = $('<li class="list-item" />').appendTo(this.$el);
+                    var li = $('<li />').addClass(csscls('list-item')).appendTo(this.$el);
                     this.get('itemRenderer')(li, data[i]);
                 }
             });
@@ -104,7 +114,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
 
         tagName: 'dl',
 
-        className: 'phpdebugbar-widgets-kvlist',
+        className: csscls('kvlist'),
 
         render: function() {
             this.bindAttr(['itemRenderer', 'data'], function() {
@@ -115,8 +125,8 @@ if (typeof(PhpDebugBar) == 'undefined') {
 
                 var self = this;
                 $.each(this.get('data'), function(key, value) {
-                    var dt = $('<dt class="key" />').appendTo(self.$el);
-                    var dd = $('<dd class="value" />').appendTo(self.$el);
+                    var dt = $('<dt />').addClass(csscls('key')).appendTo(self.$el);
+                    var dd = $('<dd />').addClass(csscls('value')).appendTo(self.$el);
                     self.get('itemRenderer')(dt, dd, key, value);
                 });
             });
@@ -148,7 +158,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
      */
     var VariableListWidget = PhpDebugBar.Widgets.VariableListWidget = KVListWidget.extend({
 
-        className: 'phpdebugbar-widgets-kvlist phpdebugbar-widgets-varlist',
+        className: csscls('kvlist varlist'),
 
         itemRenderer: function(dt, dd, key, value) {
             dt.text(key);
@@ -158,10 +168,10 @@ if (typeof(PhpDebugBar) == 'undefined') {
                 v = v.substr(0, 100) + "...";
             }
             dd.text(v).click(function() {
-                if (dd.hasClass('pretty')) {
-                    dd.text(v).removeClass('pretty');
+                if (dd.hasClass(csscls('pretty'))) {
+                    dd.text(v).removeClass(csscls('pretty'));
                 } else {
-                    dd.html(htmlize(value)).addClass('pretty');
+                    dd.html(htmlize(value)).addClass(csscls('pretty'));
                 }
             });
         }
@@ -180,7 +190,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
 
         tagName: 'iframe',
 
-        className: 'phpdebugbar-widgets-iframe',
+        className: csscls('iframe'),
 
         render: function() {
             this.$el.attr({
@@ -209,7 +219,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
      */
     var MessagesWidget = PhpDebugBar.Widgets.MessagesWidget = PhpDebugBar.Widget.extend({
 
-        className: 'phpdebugbar-widgets-messages',
+        className: csscls('messages'),
 
         render: function() {
             var self = this;
@@ -220,28 +230,28 @@ if (typeof(PhpDebugBar) == 'undefined') {
                     m = m.substr(0, 100) + "...";
                 }
 
-                var val = $('<span class="value" />').text(m).appendTo(li);
+                var val = $('<span />').addClass(csscls('value')).text(m).appendTo(li);
                 if (!value.is_string || value.message.length > 100) {
                     li.css('cursor', 'pointer').click(function() {
-                        if (val.hasClass('pretty')) {
-                            val.text(m).removeClass('pretty');
+                        if (val.hasClass(csscls('pretty'))) {
+                            val.text(m).removeClass(csscls('pretty'));
                         } else {
-                            val.html(htmlize(value.message)).addClass('pretty');
+                            val.html(htmlize(value.message)).addClass(csscls('pretty'));
                         }
                     });
                 }
 
                 if (value.label) {
-                    val.addClass(value.label);
-                    $('<span class="label" />').text(value.label).appendTo(li);
+                    val.addClass(csscls(value.label));
+                    $('<span />').addClass(csscls('label')).text(value.label).appendTo(li);
                 }
                 if (value.collector) {
-                    $('<span class="collector" />').text(value.collector).appendTo(li);
+                    $('<span />').addClass(csscls('collector')).text(value.collector).appendTo(li);
                 }
             }});
 
             this.$list.$el.appendTo(this.$el);
-            this.$toolbar = $('<div class="toolbar"><i class="icon-search"></i></div>').appendTo(this.$el);
+            this.$toolbar = $('<div><i class="icon-search"></i></div>').addClass(csscls('toolbar')).appendTo(this.$el);
 
             $('<input type="text" />')
                 .on('change', function() { self.set('search', this.value); })
@@ -249,7 +259,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
 
             this.bindAttr('data', function(data) {
                 this.set({ exclude: [], search: '' });
-                this.$toolbar.find('.filter').remove();
+                this.$toolbar.find(csscls('.filter')).remove();
 
                 var filters = [], self = this;
                 for (var i = 0; i < data.length; i++) {
@@ -257,7 +267,8 @@ if (typeof(PhpDebugBar) == 'undefined') {
                         continue;
                     }
                     filters.push(data[i].label);
-                    $('<a class="filter" href="javascript:" />')
+                    $('<a href="javascript:" />')
+                        .addClass(csscls('filter'))
                         .text(data[i].label)
                         .attr('rel', data[i].label)
                         .on('click', function() { self.onFilterClick(this); })
@@ -282,10 +293,10 @@ if (typeof(PhpDebugBar) == 'undefined') {
         },
 
         onFilterClick: function(el) {
-            $(el).toggleClass('excluded');
+            $(el).toggleClass(csscls('excluded'));
 
             var excludedLabels = [];
-            this.$toolbar.find('.filter.excluded').each(function() {
+            this.$toolbar.find(csscls('.filter') + csscls('.excluded')).each(function() {
                 excludedLabels.push(this.rel);
             });
 
@@ -306,19 +317,19 @@ if (typeof(PhpDebugBar) == 'undefined') {
 
         tagName: 'ul',
 
-        className: 'phpdebugbar-widgets-timeline',
+        className: csscls('timeline'),
 
         render: function() {
             this.bindAttr('data', function(data) {
                 this.$el.empty();
                 if (data.measures) {
                     for (var i = 0; i < data.measures.length; i++) {
-                        var li = $('<li class="measure" />');
-                        li.append($('<span class="value" />').css({
+                        var li = $('<li />').addClass(csscls('measure'));
+                        li.append($('<span />').addClass(csscls('value')).css({
                             left: Math.round(data.measures[i].relative_start * 100 / data.duration) + "%",
                             width: Math.round(data.measures[i].duration * 100 / data.duration) + "%"
                         }));
-                        li.append($('<span class="label" />').text(data.measures[i].label + " (" + data.measures[i].duration_str + ")"));
+                        li.append($('<span />').addClass(csscls('label')).text(data.measures[i].label + " (" + data.measures[i].duration_str + ")"));
                         this.$el.append(li);
                     }
                 }
@@ -337,19 +348,19 @@ if (typeof(PhpDebugBar) == 'undefined') {
      */
     var ExceptionsWidget = PhpDebugBar.Widgets.ExceptionsWidget = PhpDebugBar.Widget.extend({
 
-        className: 'phpdebugbar-widgets-exceptions',
+        className: csscls('exceptions'),
 
         render: function() {
             this.$list = new ListWidget({ itemRenderer: function(li, e) {
-                $('<span class="message" />').text(e.message).appendTo(li);
+                $('<span />').addClass(csscls('message')).text(e.message).appendTo(li);
                 if (e.file) {
-                    $('<span class="filename" />').text(e.file + "#" + e.line).appendTo(li);
+                    $('<span />').addClass(csscls('filename')).text(e.file + "#" + e.line).appendTo(li);
                 }
                 if (e.type) {
-                    $('<span class="type" />').text(e.type).appendTo(li);
+                    $('<span />').addClass(csscls('type')).text(e.type).appendTo(li);
                 }
                 if (e.surrounding_lines) {
-                    var file = $('<div class="file" />').html(htmlize(e.surrounding_lines.join(""))).appendTo(li);
+                    var file = $('<div />').addClass(csscls('file')).html(htmlize(e.surrounding_lines.join(""))).appendTo(li);
                     li.click(function() {
                         if (file.is(':visible')) {
                             file.hide();
@@ -364,7 +375,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
             this.bindAttr('data', function(data) {
                 this.$list.set('data', data);
                 if (data.length == 1) {
-                    this.$list.$el.children().first().find('.file').show();
+                    this.$list.$el.children().first().find(csscls('.file')).show();
                 }
             });
 
@@ -382,35 +393,34 @@ if (typeof(PhpDebugBar) == 'undefined') {
      */
     var SQLQueriesWidget = PhpDebugBar.Widgets.SQLQueriesWidget = PhpDebugBar.Widget.extend({
 
-        className: 'phpdebugbar-widgets-sqlqueries',
+        className: csscls('sqlqueries'),
 
         render: function() {
-            this.$status = $('<div class="status" />').appendTo(this.$el);
+            this.$status = $('<div />').addClass(csscls('status')).appendTo(this.$el);
 
             this.$list = new ListWidget({ itemRenderer: function(li, stmt) {
-                $('<span class="sql" />').text(stmt.sql).appendTo(li);
+                $('<span />').addClass(csscls('sql')).text(stmt.sql).appendTo(li);
                 if (stmt.duration_str) {
-                    $('<span class="duration" title="Duration" />').text(stmt.duration_str).appendTo(li);
+                    $('<span title="Duration" />').addClass(csscls('duration')).text(stmt.duration_str).appendTo(li);
                 }
                 if (stmt.memory_str) {
-                    $('<span class="memory" title="Peak memory usage" />').text(stmt.memory_str).appendTo(li);
+                    $('<span title="Peak memory usage" />').addClass(csscls('memory')).text(stmt.memory_str).appendTo(li);
                 }
                 if (typeof(stmt.is_success) != 'undefined' && !stmt.is_success) {
-                    li.addClass('error');
-                    li.append($('<span class="error" />').text("[" + stmt.error_code + "] " + stmt.error_message));
+                    li.addClass(csscls('error'));
+                    li.append($('<span />').addClass(csscls('error')).text("[" + stmt.error_code + "] " + stmt.error_message));
                 } else if (typeof(stmt.row_count) != 'undefined') {
-                    $('<span class="row-count" title="Row count" />').text(stmt.row_count).appendTo(li);
+                    $('<span title="Row count" />').addClass(csscls('row-count')).text(stmt.row_count).appendTo(li);
                 }
                 if (typeof(stmt.stmt_id) != 'undefined' && stmt.stmt_id) {
-                    $('<span class="stmt-id" title="Prepared statement ID" />').text(stmt.stmt_id).appendTo(li);
+                    $('<span title="Prepared statement ID" />').addClass(csscls('stmt-id')).text(stmt.stmt_id).appendTo(li);
                 }
                 if (stmt.params && !$.isEmptyObject(stmt.params)) {
-                    var table = '<table class="params"><tr><th colspan="2">Params</th></tr>';
+                    var table = $('<table><tr><th colspan="2">Params</th></tr></table>').addClass(csscls('params')).appendTo(li);
                     for (var key in stmt.params) {
-                        table += '<tr><td class="name">' + key + '</td><td class="value">' + stmt.params[key] + '</td></tr>';
+                        table.append('<tr><td class="' + csscls('name') + '">' + key + '</td><td class="' + csscls('value') + 
+                                     '">' + stmt.params[key] + '</td></tr>');
                     }
-                    table += '</table>';
-                    table = $(table).appendTo(li);
                     li.css('cursor', 'pointer').click(function() {
                         if (table.is(':visible')) {
                             table.hide();
@@ -431,10 +441,10 @@ if (typeof(PhpDebugBar) == 'undefined') {
                     t.append(", " + data.nb_failed_statements + " of which failed");
                 }
                 if (data.accumulated_duration_str) {
-                    this.$status.append($('<span class="duration" title="Accumulated duration" />').text(data.accumulated_duration_str));
+                    this.$status.append($('<span title="Accumulated duration" />').addClass(csscls('duration')).text(data.accumulated_duration_str));
                 }
                 if (data.peak_memory_usage_str) {
-                    this.$status.append($('<span class="memory" title="Peak memory usage" />').text(data.peak_memory_usage_str));
+                    this.$status.append($('<span title="Peak memory usage" />').addClass(csscls('memory')).text(data.peak_memory_usage_str));
                 }
             });
         }
@@ -451,15 +461,15 @@ if (typeof(PhpDebugBar) == 'undefined') {
      */
     var TemplatesWidget = PhpDebugBar.Widgets.TemplatesWidget = PhpDebugBar.Widget.extend({
 
-        className: 'phpdebugbar-widgets-templates',
+        className: csscls('templates'),
 
         render: function() {
-            this.$status = $('<div class="status" />').appendTo(this.$el);
+            this.$status = $('<div />').addClass(csscls('status')).appendTo(this.$el);
 
             this.$list = new ListWidget({ itemRenderer: function(li, tpl) {
-                $('<span class="name" />').text(tpl.name).appendTo(li);
+                $('<span />').addClass(csscls('name')).text(tpl.name).appendTo(li);
                 if (tpl.render_time_str) {
-                    $('<span class="render_time" title="Render time" />').text(tpl.render_time_str).appendTo(li);
+                    $('<span title="Render time" />').addClass(csscls('render_time')).text(tpl.render_time_str).appendTo(li);
                 }
             }});
             this.$list.$el.appendTo(this.$el);
@@ -468,7 +478,7 @@ if (typeof(PhpDebugBar) == 'undefined') {
                 this.$list.set('data', data.templates);
                 this.$status.empty().append($('<span />').text(data.templates.length + " templates were rendered"));
                 if (data.accumulated_render_time_str) {
-                    this.$status.append($('<span class="render_time" title="Accumulated render time" />').text(data.accumulated_render_time_str));
+                    this.$status.append($('<span title="Accumulated render time" />').addClass(csscls('render_time')).text(data.accumulated_render_time_str));
                 }
             });
         }
@@ -485,14 +495,14 @@ if (typeof(PhpDebugBar) == 'undefined') {
      */
     var MailsWidget = PhpDebugBar.Widgets.MailsWidget = PhpDebugBar.Widget.extend({
 
-        className: 'phpdebugbar-widgets-mails',
+        className: csscls('mails'),
 
         render: function() {
             this.$list = new ListWidget({ itemRenderer: function(li, mail) {
-                $('<span class="subject" />').text(mail.subject).appendTo(li);
-                $('<span class="to" />').text(mail.to).appendTo(li);
+                $('<span />').addClass(csscls('subject')).text(mail.subject).appendTo(li);
+                $('<span />').addClass(csscls('to')).text(mail.to).appendTo(li);
                 if (mail.headers) {
-                    var headers = $('<pre class="headers" />').appendTo(li);
+                    var headers = $('<pre />').addClass(csscls('headers')).appendTo(li);
                     $('<code />').text(mail.headers).appendTo(headers);
                     li.click(function() {
                         if (headers.is(':visible')) {